Move cnl::from_rep into a bits/type_traits/
