+++ include/cnl/bits/num_traits/from_rep.h
@@ -1,0 +1,40 @@
+
+//          Copyright John McFarlane 2018.
+// Distributed under the Boost Software License, Version 1.0.
+//    (See accompanying file ../LICENSE_1_0.txt or copy at
+//          http://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef CNL_BITS_NUM_TRAITS_FROM_REP
+#define CNL_BITS_NUM_TRAITS_FROM_REP
+
+#include <cnl/bits/type_traits/is_integral.h>
+#include <cnl/bits/type_traits.h>
+
+namespace cnl {
+    /// \brief generic function object that returns the number encapsulating a given value
+    ///
+    /// \tparam Number archetype for the encapsulating type
+    ///
+    /// \note Rather than returning Number, invocation may return an alternative
+    /// template instantiation based on input parameter.
+    /// \sa to_rep, from_value
+    template<class Number, class Enable = void>
+    struct from_rep;
+
+    /// \brief Specialization of \ref from_rep for integer types
+    ///
+    /// \tparam Number fundamental integer type to return
+    ///
+    /// \note This specialization *does* return integers of type, \c Number
+    /// \sa to_rep, from_value
+    template<class Number>
+    struct from_rep<Number, _impl::enable_if_t<cnl::_impl::is_integral<Number>::value>> {
+        template<class Rep>
+        constexpr Number operator()(Rep const& rep) const {
+            // by default, a number type's rep type is the number type itself
+            return static_cast<Number>(rep);
+        }
+    };
+}
+
+#endif  // CNL_BITS_NUM_TRAITS_FROM_REP
